requirements.txt
# This requirements file contains the requirements that are
# Dependencies for testing
ray[serve]
# Dependencies for loading & serving the model:

colpali-engine>=0.3.0,<0.4.0

# These are useful for loading most Huggingface models:
torch
transformers

# # You'll probably need Pillow & python-multipart for models that take images as input:
Pillow
python-multipart
